Revert "rename operations -> drawingOperations"
This reverts commit b4043c6871a6b8511f4febdbb2746a62a074da1f.
