+++ src/api/index.ts
@@ -4,7 +4,7 @@ export * from './colors';
 export * from './errors';
 export * from './image';
 export * from './objects';
-export * from './drawingOperations';
+export * from './operations';
 export * from './operators';
 export * from './rotations';
 export * from './sizes';
